requirements.txt
amqp==5.3.1
billiard==4.2.1
celery==5.5.3
certifi>=2025.11.12
chardet>=5.1.0
Django==5.2.9
django-celery-results>=2.5.0
django-ranged-response>=0.2.0
django-simple-captcha>=0.6.0
django-widget-tweaks>=1.5.0
django-debug-toolbar>=4.0.0
django-background-tasks>=1.2.8
djutils>=0.3.2
gunicorn>=21.0.0
idna>=3.5
kombu>=5.5.2,<5.6
Pillow>=10.0.0
psycopg2>=2.9.10
python-bitcoinrpc>=1.0
pytz>=2025.2
requests>=2.31.0
six>=1.16.0
urllib3>=2.0.4
vine>=5.0.0
gnupg>=2.3.1
django_countries>=8.0.0
cryptowatch-client>=0.1.4
django-star-ratings>=0.9.0
pymysql>=1.1.0
django-avatar>=6.0.0
